% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{addcount}
\alias{addcount}
\title{addcount - adds a count of years and avC to a data.frame}
\usage{
addcount(indat, varid = "Vessel", group = "Year", catch = "catch_kg")
}
\arguments{
\item{indat}{- a data.frame containing at least columns named 'catch_kg',
'Vessel' (or in fact 'varid'), and 'Year'.}

\item{varid}{the variable in the data.frame for which the count and
avearge catch per year is to be estimated. Defaults to 'Vessel'
ready for the SESSF but could be diver_id for abalone, or any other
factor of interest.}

\item{group}{the first grouping variable, typically Vessel in the SESSF}

\item{catch}{the variable name for the catch, catch_kg in the SESSF}
}
\value{
a data.frame made of the input data.frame plus two new columns
}
\description{
addcount - given an input data.frame containing catch_kg,
   Vessel, and Year as variables this function first sums the catch by
   vessel, by year, then counts each vessels occurrence across the
   years, and each vessels catch. Then stepping through each vessel it
   populates two new colummns named 'count' and 'avC'.
}
