% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{countgtzero}
\alias{countgtzero}
\title{countgtzero used in apply to count the number of zeros in a vector}
\usage{
countgtzero(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of zeros
}
\description{
countgtzero used in apply to count the number of zeros in a vector
}
\examples{
x <- rnorm(30,mean=1,sd=1)
countgtzero(x)
cat(30 - countgtzero(x),"out of ",30, " are less than 0  \\n")
y <- matrix(x,nrow=6,ncol=5)
apply(y,1,countgtzero)
}
