% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{getsps}
\alias{getsps}
\title{getsps reads in the species catch and effort data}
\usage{
getsps(filename, datadir, depthclass = 25, txtout = FALSE)
}
\arguments{
\item{filename}{the names of the csv file to be read}

\item{datadir}{the path to the directory containing the csv files}

\item{depthclass}{the class width of the depth categories, defaults to 25}

\item{txtout}{report the filename to the screen? default = FALSE}
}
\value{
invisibly returns the sps data.frame ready for selection
}
\description{
getsps reads in the species catch and effort data that was
    previously extracted from the Oracle data base.
}
\examples{
\dontrun{
print("need an example that will work")
}
}
