% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{makecategorical}
\alias{makecategorical}
\title{makecategorical - converts given variables into categorical factors}
\usage{
makecategorical(labelModel, indat)
}
\arguments{
\item{labelModel}{is the set of variables from the data.frame that are to
be converted into categorical factors.}

\item{indat}{is the data.frame that is to be analysed in the standardization}
}
\value{
a copy of the database with the selected variables converted into
    factors
}
\description{
Given a list of variables, as character strings, this function
  converts each variable into a factor after checking no mistake has been
  made with the spelling of the variable name. A copy of the data is made
  so that the non factored variables can be retained.
}
\examples{
\dontrun{
data(sps)
labelM <- c("Year","Vessel","Month")
sps1 <- makecategorical(labelM,sps)
}
}
