% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{makedatasum}
\alias{makedatasum}
\title{makedatasum generates a table summarizing information about data}
\usage{
makedatasum(InData, InData1, out)
}
\arguments{
\item{InData}{the unselected species data, generally sps}

\item{InData1}{the data.frame after selection has been applied, sps1}

\item{out}{the output from standLM or dosingle, usually out}
}
\value{
a matrix of each summary variable by year
}
\description{
makedatasum uses the two input data.frames to generates a
    table summarizing information about data, this includes the Total
    Catch of the species, number of records used, the catch used in the
    analyses, the number of Vessels, the GeoMean, the optimum standardized
    mean estimates, their StDev, the total catch <30Kg, and the
    proportion of catch <30Kg.
}
\examples{
\dontrun{
data(sps)
sps$CE <- NA
sps$LnCE <- NA
pick <- which((sps$catch_kg > 0) & (sps$Effort > 0))
sps$CE[pick] <- sps$catch_kg[pick]/sps$Effort[pick]
sps$LnCE[pick] <- log(sps$CE[pick])
answer <- selectdata(sps,dependent="LnCE",Ldepth=0,Udepth=600,
startyear=1986,finalyear=2015,zones=c(10,20,30),method=c("TW","TDO"),
fishery="SET")
sps1 <- answer$sps1
makedatasum(sps,sps1)
}
}
