% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{plotfishery}
\alias{plotfishery}
\title{plotfishery plots an array of data for the fishery.}
\usage{
plotfishery(InData2, Datain, spsname = "", Ldep = 0, Udep = 1000,
  zones = "Zone", legpos = "L", legval = 0.6)
}
\arguments{
\item{InData2}{is he selected data from selectdata, usually sps1}

\item{Datain}{is the datasum from makedatasum}

\item{spsname}{is the original splabel, usually splabel}

\item{Ldep}{lowest selected depth usually Ldepth, defaults to 0}

\item{Udep}{deepest selected depth, usually Udepth, defaults to 1000}

\item{zones}{defaults to 'Zone' to represent the SESSF zones, but if the
fishery relates to the Orange Roughy zones then use 'ORzone', and if it
relates to the Shark fishery use "SharkRegion'}

\item{legpos}{defaults to "L", which will put the legend on the left of the
plot. Any other entry, such as "R" will place it on the right hand side}

\item{legval}{defaults to 0.6, which should work for most plots unless the
x-axis does not start at 0, in which case a larger value may be required.}
}
\value{
currently nothing but this does generate a 3 x 2 plot
}
\description{
plotfishery plots an array of data for the fishery. Including
    the number of records by depth, the catch-by-cept by zone, the number
    of vessels by year, the number of records by year, the total catch by
    all methods, gears, and zones by year, combined with the selected
    catches-by year, then the selected catches-by-year to increase the
    y-axis scale if necessary to calrify the total catch < 30kg by year.
}
\examples{
\dontrun{
print("still to develop a full example")
}
}
