% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplots.R
\name{plotstandFY}
\alias{plotstandFY}
\title{plotstandFY - optimum model & Geometric mean for fishing years}
\usage{
plotstandFY(stnd, yrtick = 4, bars = FALSE, P = 95, lstyr = NA,
  alterpar = TRUE)
}
\arguments{
\item{stnd}{is the list output from standLM}

\item{yrtick}{defaults to 4 denoting the gap between years on the
x-axis}

\item{bars}{is a logical T or F determining whether to put confidence bounds
around each estimate; defaults to FALSE}

\item{P}{is the percentile used for the log-normal confidence bounds, if
they are plotted; defaults to 95.}

\item{lstyr}{is last years time-series of the optimal model. This can be
plotted on top of the current year's optimum to chack on repeatability.}

\item{alterpar}{is a logical that, if TRUE, which is the default state, it
will prevent this function from altering the par settings.
If FALSE this will allow for additions to be made to the resulting plot.}
}
\value{
a plot of the model with the smallest AIC (solid line) and the
   geometric mean (model 1, always = LnCE ~ Year, the dashed line). 'Year'
   could be some other time step.
}
\description{
plotstandFY - optimum model & Geometric mean for fishing years.
   Standardizations from standLM and standLnCE can be used.
   Has options that allow for log-normls P% intervals around each time
   period's parameter estimate. Also can rescale the graph to have an average
   the same as the geometric mean average of the original time series of data.
}
\examples{
\dontrun{
data(sps)
splabel <- "Species2"
labelM <- c("Year","Vessel","Month")
sps1 <- makecategorical(labelM,sps)
mods <- makemodels(labelM)
out <- standLM(mods,sps1,splabel)
plotstandFY(out, bars=TRUE, P=90)
plotstandFY(out)
}
}
