% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{selectdata}
\alias{selectdata}
\title{selectdata - select records from the sessf data}
\usage{
selectdata(indat, dependent = "LnCE", Ldepth = 0, Udepth = 1000,
  startyear = 0, finalyear = 3000, inzones = NA, ORzone = NA,
  sharkzone = NA, method = NA, fishery = NA)
}
\arguments{
\item{indat}{the data set to be filtered - usually sps}

\item{dependent}{- the dependent variable in the standardization,
defaults to LnCE}

\item{Ldepth}{the lower depth bound - defaults to 0}

\item{Udepth}{the upper depth bound, defaults to 1000}

\item{startyear}{the first year for inclusion, defaults to 0, which
implies the use of all years present}

\item{finalyear}{the final year for inclusion, defaults to 0}

\item{inzones}{the sessf zones to be considered, defaults to NA,
which implies either ORzone or sharkzone will be used}

\item{ORzone}{if the species uses the Orange Roughy zones rather than the
SESSF zones then use this input variable; defaults to NA}

\item{sharkzone}{if the species uses the Shark Regions rather than the
SESSF zones then use this input variable; defaults to NA}

\item{method}{the fishing method to be considered, defaults to NA,
which implies all methods to be included}

\item{fishery}{the three letter acronym for eahc fishery to be
considered, defaults to NA, which means include all.}
}
\value{
a list made up of the filtered data set named 'sps1', and
   a matrix containing the effect of the filtering on the numbers of
   observations; named 'nobs'.
}
\description{
selectdata - select records from the sessf data based upon
   whether ther edependent variable is empty or not, depth range, years,
   SESSF zones, fishing method, and fishery. If any are omitted then
   all records will be used
}
\examples{
\dontrun{
data(sps)
sort(unique(sps$Year))
print(range(sps$Depth))
sps$CE <- NA
sps$LnCE <- NA
pick <- which((sps$catch_kg > 0) & (sps$Effort > 0))
sps$CE[pick] <- sps$catch_kg[pick]/sps$Effort[pick]
sps$LnCE[pick] <- log(sps$CE[pick])
out <- selectdata(sps)
sps1 <- out$sps1
nobs <- out$nobs
nobs
}
}
