% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_doc.R
\docType{data}
\name{sps}
\alias{sps}
\title{sps - a 11603 x 10 data.frame for testing CPUE functions}
\format{A data frame with 11603 records x 10 variables:
\itemize{
  \item Year - The year in which fishing takes place
  \item Month - The month in which fishing took place
  \item Vessel - a code uniquely identifying vessels through time
  \item catch_kg - the retained catch of the target species in kg
  \item Long - the longitude of the start of the trawl shot
  \item Lat - the latitude of the start of the trawl shot
  \item Depth - the average depth of trawling in meters
  \item DayNight - a code denoting the daynight status D = day, N = night,
     M = mixed, and U = unknown
  \item Effort - the hours trawled
  \item Zone the xone in which trawling occurs 1 - 3
}}
\description{
sps - a 11603 x 10 data.frame for testing CPUE functions
   containing simulated trawl shot CPUE data for the years 2003 - 2014,
   including details of year, month, vessel, catch_kg, longitude,
   latitude, the depth trawled, a daynight identifier, the effort
   in hours, and the zone, the CPUE, and the natural log(CPUE) need
   to be added.
}
