% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuefuncs.r
\name{standLM}
\alias{standLM}
\title{standLM Conduct a standarization on log-transformed CPUE using lm}
\usage{
standLM(inmods, indat, inlab = "", console = TRUE)
}
\arguments{
\item{inmods}{is the list of models to be analysed along with their
respective labels - can be generated using makemodels}

\item{indat}{is the data.frame containing the raw data, which must contain
columns with the same names as the factors being included in the models.}

\item{inlab}{is an optional label that is added to text and plotted outputs,
which defaults to the empty string}

\item{console}{defaults to TRUE; if TRUE write each model to the screen
as each model runs, which can mess up auto-documenting documents.}
}
\value{
a list containing the standardization of class 'CEout'
}
\description{
Conduct a standarization on log-transformed CPUE data using lm.
  A number of predefined models are entered (use makemodels) with the raw
  data and a label. The optimum model is taken to be the one with the
  smallest AIC, though usually (not always) this is the same as with the 
  largest adj-r2.
}
\examples{
\dontrun{
data(ab)
splabel = "Abalone"
print(tapply(ab$catch,list(ab$year,ab$block),sum,na.rm=TRUE)/1000)
labelM <- c("year","diverID","month","block","boatID")
ab1 <- makecategorical(labelM,ab)
mods <- makemodels(labelM)
out <- standLM(mods,ab1,splabel)
round(out$Results,4)
round(out$WhichM,4)
plotstand(out,bars=TRUE)
}
}
